+++ packages/start-plugin-core/src/dev-server-plugin/extract-html-scripts.ts
