+++ packages/start-plugin-core/src/load-env-plugin/plugin.ts
@@ -1,5 +1,5 @@
 import * as vite from 'vite'
-import type { TanStackStartOutputConfig } from '../../plugin'
+import type { TanStackStartOutputConfig } from '../plugin'
 
 export function loadEnvPlugin(
   startOpts: TanStackStartOutputConfig,
